+++ .bak/Makefile.bak
@@ -28,7 +28,6 @@ SRC = \
   src/token/tokenize_operator.c \
   src/token/tokenize_frame.c \
   src/token/utils.c \
-  src/token/tokenize_quote.c \
   src/token/ms_tokenize.c 
 ### END AUTO GENERATED FILES ###
