+++ Makefile
@@ -28,7 +28,12 @@ SRC = \
   src/token/tokenize_operator.c \
   src/token/tokenize_frame.c \
   src/token/utils.c \
-  src/token/ms_tokenize.c 
+  src/token/ms_tokenize.c \
+  src/ast/print_ast.c \
+  src/ast/utils.c \
+  src/ast/parse.c \
+  src/ast/ms_parse_tokens.c \
+  src/ast/create_node.c 
 ### END AUTO GENERATED FILES ###
 
 # **************************************************************************** #
